+++ infra/terraform/infra/containers.tf
@@ -4,7 +4,7 @@ resource "null_resource" "deploy_containers" {
   depends_on = [module.docker]
 
   provisioner "local-exec" {
-    command = "ansible-playbook playbooks/gitlab.yml"
+    command = "ansible-playbook --vault-password-file ${var.ansible_vault_key} playbooks/gitlab.yml"
     working_dir = "../../ansible"
   }
 }